Add RoomAllocationProblem
Generate a solution with a random room.
Verify if the slots needed for the lecture are available in such random room.
If so, allocates it in the solution. If not, find a new random room by an index Int.

+++ src/main/kotlin/com/iscte/ads/schedulegen/study/NSGAIIStudy1.kt
@@ -1,6 +1,7 @@
 package com.iscte.ads.schedulegen.study
 
 import com.iscte.ads.schedulegen.datamapping.ObjectConverter
+import com.iscte.ads.schedulegen.room.Room
 import com.iscte.ads.schedulegen.room.RoomsGatewayImplementation
 import com.iscte.ads.schedulegen.schedule.StudentClass
 import com.iscte.ads.schedulegen.schedule.TimeSlot
@@ -12,20 +13,19 @@ import java.io.File
 import java.io.IOException
 import java.nio.file.Files
 import java.nio.file.Paths
-import java.util.Arrays;
+import java.util.Arrays
 
-import org.apache.commons.io.FileUtils;
-import org.uma.jmetal.algorithm.multiobjective.nsgaii.NSGAIIBuilder;
-import org.uma.jmetal.lab.experiment.Experiment;
-import org.uma.jmetal.lab.experiment.ExperimentBuilder;
+import org.apache.commons.io.FileUtils
+import org.uma.jmetal.algorithm.multiobjective.nsgaii.NSGAIIBuilder
+import org.uma.jmetal.lab.experiment.Experiment
+import org.uma.jmetal.lab.experiment.ExperimentBuilder
 import org.uma.jmetal.lab.experiment.component.impl.ComputeQualityIndicators
-import org.uma.jmetal.lab.experiment.component.impl.ExecuteAlgorithms;
+import org.uma.jmetal.lab.experiment.component.impl.ExecuteAlgorithms
 import org.uma.jmetal.lab.experiment.component.impl.GenerateReferenceParetoFront
-import org.uma.jmetal.lab.experiment.component.impl.GenerateReferenceParetoSetAndFrontFromDoubleSolutions
-import org.uma.jmetal.lab.experiment.util.ExperimentAlgorithm;
-import org.uma.jmetal.lab.experiment.util.ExperimentProblem;
-import org.uma.jmetal.qualityindicator.impl.NormalizedHypervolume;
-import org.uma.jmetal.qualityindicator.impl.Spread;
+import org.uma.jmetal.lab.experiment.util.ExperimentAlgorithm
+import org.uma.jmetal.lab.experiment.util.ExperimentProblem
+import org.uma.jmetal.qualityindicator.impl.NormalizedHypervolume
+import org.uma.jmetal.qualityindicator.impl.Spread
 
 
 object NSGAIIStudy2 {
@@ -52,13 +52,23 @@ object NSGAIIStudy2 {
 
         val timeSlots: List<TimeSlot> = gateway.convertFromTimeSlotsCsv(timeSlotsCsv)
 
+
+        val roomsCsv = NSGAIIStudy::class.java.getResource("/rooms.csv")?.readText()
+        if (roomsCsv == null) {
+            println("rooms file content is null - returning")
+            return
+        }
+
+        val rooms: List<Room> = gateway.convertFromRoomsCsv(roomsCsv)
+
         val scheduleProblem = ScheduleGenProblem(lectures, timeSlots)
+        val roomAllocationProblem = RoomAllocationProblem(lectures, timeSlots, rooms)
 
         val experimentBaseDirectory = "experimentBaseDirectory"
 
         // Apagar os dados da simulação anterior, não é feito pelo jMetal
         FileUtils.deleteDirectory(File(experimentBaseDirectory))
-        val problemList: List<ExperimentProblem<IntegerSolution>> = listOf(ExperimentProblem(scheduleProblem))
+        val problemList: List<ExperimentProblem<IntegerSolution>> = listOf(ExperimentProblem(scheduleProblem), ExperimentProblem(roomAllocationProblem))
 
         val algorithmList: MutableList<ExperimentAlgorithm<IntegerSolution, List<IntegerSolution>>> = configureAlgorithmList(problemList)
         val experiment: Experiment<IntegerSolution, List<IntegerSolution>> =
